created local Docker dev environment including database copy

+++ migration_scripts/003_coordinates_split.sql
@@ -1,2 +1,2 @@
-ALTER TABLE melding ADD COLUMN locatie_lat REAL;
-ALTER TABLE melding ADD COLUMN locatie_lon REAL;
+ALTER TABLE melding ADD COLUMN locatie_lat NUMERIC;
+ALTER TABLE melding ADD COLUMN locatie_lon NUMERIC;
